[REFACTOR] Move base64url helpers to helpers
